Add `AnalysisTask`.
`AnalysisTask` allows a task to specify result files that need to be
copied to the worker node before it begins analysis. The framework will
automatically deploy and clean up these files.

Resource deployment is done synchronously by the management node in case
multiple tasks are launched that require large data files. Cleanup is
performed on the worker node after the task has executed.