Fix bug when selecting worker node for a task.
When selecting a worker for a task, there was no check in place to
ensure that the worker had enough open slots to perform the task. For
example, a worker with 1 open slot could be assigned a task that
requires 2 slots.

This commit addresses this issue and cleans up the readability of the
code.