Revise task results.
Update result interface. Callers now specify all the pieces of the path
of the result file. Callers can also optionally specify the store
directory.

Also add a second type of results: logs. Logs are not necessary for
recovery. When storing results, missing logs are ignored and do not
constitute an error.

Finally, the hostname of the worker is stored by all messages. Tasks can
use this to internally manage the 'host' parameter when storing results.
This recudes the ammount of information stored for each result, and also
lets results be added before a task is deployed (although '_root'
becomes invalid once deployed).